better tables

+++ branches/staticJS/webEdition/liveUpdate/sqldumps/tblNewsletterConfirm.sql
@@ -7,7 +7,7 @@ CREATE TABLE ###TBLPREFIX###tblNewsletterConfirm (
   subscribe_firstname varchar(255) NOT NULL default '',
   subscribe_lastname varchar(255) NOT NULL default '',
   lists text NOT NULL,
-  expires bigint(20) unsigned NOT NULL default '0',
+  expires int(11) unsigned NOT NULL default '0',
   PRIMARY KEY (confirmID(30)),
   KEY expires (expires),
   KEY subscribe (subscribe_mail(50),confirmID(30))
